% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluating.R
\name{evaluating}
\alias{evaluating}
\title{Set context for evaluation task}
\usage{
evaluating(model, task, ...)
}
\arguments{
\item{model}{A single string describing the model used to generate output.}

\item{task}{A single string describing the task being evaluated on.}
}
\description{
This helper is analogous to \code{testthat::context()} and sets metadata
for the test for logging.
}
