% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect.R
\name{expect_r_code}
\alias{expect_r_code}
\title{Check if input is syntactically valid R code}
\usage{
expect_r_code(object)
}
\arguments{
\item{object}{A single string, possibly containing parse-able R code.}
}
\value{
Returns nothing but throws an error if the input is not valid R code.
}
\description{
Check if input is syntactically valid R code
}
\details{
This helper does not attempt to evaluate the code in the provided string.
Instead, it uses \code{\link[=parse]{parse()}} to determine whether the string
contains valid R syntax.

This expectation will fail if \code{object} contains any exposition surrounding
the code, including if code is surrounded by backticks.
}
