% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade-output.R
\name{grade_output}
\alias{grade_output}
\title{Grade model outputs}
\usage{
grade_output(target)
}
\arguments{
\item{target}{A single string describing the desired output. Set to \code{NULL}
to omit desired output, though this is discouraged unless you are
grading the output yourself.}
}
\description{
Grade model outputs
}
