% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade-queue.R
\name{grade_queue}
\alias{grade_queue}
\title{Grade an evaluation data frame}
\usage{
grade_queue(x, judges, type = c("pairwise", "score"))
}
\arguments{
\item{judges}{A \code{\link[=judges]{judges()}} object or \code{NULL} to grade the evals yourself.}

\item{type}{One of \code{"pairwise"} or \code{"score"}. See details for more.}
}
\description{
If you set the \code{judges} argument to a non-NULL value in \code{\link[=evaluate]{evaluate()}},
this function will be evoked automatically.
}
\examples{
library(ellmer)

ggplot2 <- evaluate(
  "tests/evalthat/test-ggplot2-graded.R",
    across = tibble(chat = c(
        chat_openai(model = "gpt-4o-mini", echo = FALSE),
        chat_claude(model = "claude-3-5-sonnet-latest", echo = FALSE),
        chat_ollama(model = "qwen2.5-coder:14b", echo = FALSE))
      ),
      repeats = 2
    )

ggplot2


ggplot2_graded
}
