% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io}
\alias{io}
\alias{input}
\alias{output}
\title{Flag model inputs and outputs}
\usage{
input(x)

output(x)
}
\arguments{
\item{x}{A character string representing either the model input or output.}
}
\value{
\code{x}, as provided. Called for its side effect.
}
\description{
When evaluating model input in evaluations, flag the input with \code{input()}
and the output with \code{output()}, allowing for persistent logging of
responses.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
test_that("the model works", {
  input <- input("some input")
  output <- output(some_function_call_generating_response(input))
})
\dontshow{\}) # examplesIf}
}
