% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge.R
\name{judges}
\alias{judges}
\title{Define judge models}
\usage{
judges(...)
}
\arguments{
\item{...}{Named (optionally) \code{\link[ellmer:Chat]{ellmer::Chat()}} objects, e.g. the output of
\code{\link[ellmer:chat_openai]{ellmer::chat_openai()}} or \code{\link[ellmer:chat_claude]{ellmer::chat_claude()}}.}
}
\description{
Judge models allow for automated grading of model output
(with \code{grade_model()}) by passing responses to other LLM judges.

To learn more about how evalthat makes use of judge models, see
\code{vignette("Model grading", package = "evalthat")}.
}
\examples{
library(ellmer)

j <- judges(
  gpt4o = chat_openai(model = "gpt-4o"),
  claude = chat_claude()
)

j

# doesn't necessarily need to be named:
judges(chat_openai(model = "gpt-4o"), chat_claude())
}
