% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results_read}
\alias{results_read}
\title{Read eval results}
\usage{
results_read(dir = NULL)
}
\arguments{
\item{dir}{Optional. A single string specifying a subdirectory of
\verb{tests/evalthat/_results}. If \code{NULL}, will read each subdirectory and
row-bind the results.}
}
\value{
A tibble containing the combined results from result files in the
specified directory.
}
\description{
Read eval results
}
