% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalthat-package.R
\docType{package}
\name{evalthat-package}
\alias{evalthat}
\alias{evalthat-package}
\title{evalthat: LLM Evaluation in R}
\description{
Provides testthat-style utilities for evaluating large language models in R. Score models with non-deterministic output using a unit test interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/evalthat}
  \item \url{https://simonpcouch.github.io/evalthat/}
  \item Report bugs at \url{https://github.com/simonpcouch/evalthat/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

}
\keyword{internal}
