% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalthat-package.R
\docType{package}
\name{evalthat-package}
\alias{evalthat}
\alias{evalthat-package}
\title{evalthat: Testthat-Style LLM Evaluation}
\description{
Testthat-Style LLM Evaluation
}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

}
\keyword{internal}
